pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "epaper-emulator"
version = "1.0.0"
description = "Emulator for Waveshare E-Paper Display (EPD) screens"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Benjamin Burzan" },
]
keywords = ["e-paper", "epd", "waveshare", "emulator", "display"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "Pillow>=10.0.0",
    "Flask>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "flake8>=6.0",
]

[project.urls]
Homepage = "https://github.com/benjaminburzan/E-Paper-Emulator"
Repository = "https://github.com/benjaminburzan/E-Paper-Emulator"
Issues = "https://github.com/benjaminburzan/E-Paper-Emulator/issues"

[tool.setuptools.packages.find]
include = ["epaper_emulator*"]

[tool.setuptools.package-data]
epaper_emulator = ["config/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
